requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
dill==0.3.8
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
fuzzywuzzy==0.18.0
huggingface-hub==0.20.3
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
multiprocess==0.70.16
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyparsing==3.1.1
PyTDC==0.4.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rdkit==2023.9.4
rdkit-pypi==2022.9.5
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
